Core: Add cast to first argument of isalnum
 * lasso/xml/xml.c:
 * lasso/xml/tools.c:
   isalnum takes a int as first arg.

+++ lasso/xml/xml.c
@@ -1592,7 +1592,7 @@ is_base64(const char *message)
 	const char *c;
 
 	c = message;
-	while (*c != 0 && (isalnum(*c) || *c == '+' || *c == '/' || *c == '\n' || *c == '\r')) c++;
+	while (*c != 0 && (isalnum((int)*c) || *c == '+' || *c == '/' || *c == '\n' || *c == '\r')) c++;
 	while (*c == '=' || *c == '\n' || *c == '\r') c++; /* trailing = */
 
 	if (*c == 0)
